feat: add background correction in quickview
Since Shape-In now also computes the rolling mean background image,
this feature allows to switch on/off a background correction using
this rolling mean. It can be activated by a checkbox in the quickview
window. For datasets that do not contain the feature "image_bg" this
checkbox will not be visible.

Also added pytest-qt tests for this new feature.